calendar improvements
