+++ src/App.jsx
@@ -16,6 +16,7 @@ function App() {
         <Navbar />
         <Routes>
           <Route path="/login" element={<Login />}></Route>
+          <Route path="/schedulet" element={<SchedulePage />} />
           <Route element={<ProtectedRoute />}>
             <Route path="/" element={<SchedulePage />} />
             <Route path="/schedule" element={<SchedulePage />} />
